+++ ofdft_ml/model/__init__.py
@@ -1,2 +1,0 @@
-from .GP_model import GP_model
-from .GPGD_model import GPGD_model